% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bench.R
\name{bench}
\alias{bench}
\title{Benchmark a function across different package versions}
\usage{
bench(
  fn,
  ...,
  pkgs,
  args = list(),
  libpath = .libPaths(),
  args_pak = list(),
  args_callr = list()
)
}
\arguments{
\item{fn}{\verb{[function]}

A function to evaluate. The function is passed along to \code{\link[callr:r]{callr::r()}}, so
it is evaluated in a fresh R session and must be self-contained.

Read the \code{func} docs of \code{\link[callr:r]{callr::r()}} for the full set of restrictions on
\code{fn}.

\code{fn} is converted to a function with \code{\link[rlang:as_function]{rlang::as_function()}}, so it can be
a lambda function.}

\item{...}{These dots are for future extensions and must be empty.}

\item{pkgs}{\verb{[character]}

A character vector of package names or remote package specifications to
install and run \code{fn} against. Passed along to \code{\link[pak:pkg_install]{pak::pkg_install()}}.}

\item{args}{\verb{[list]}

An optional list of arguments to pass to the function.}

\item{libpath}{\verb{[character]}

The base library path to use.}

\item{args_pak}{\verb{[named list]}

A named list of arguments to pass on to \code{\link[pak:pkg_install]{pak::pkg_install()}}.

Can't include:
\itemize{
\item \code{pkg}
\item \code{lib}
\item \code{ask}
}}

\item{args_callr}{\verb{[named list]}

A named list of arguments to pass on to \code{\link[callr:r]{callr::r()}}.

Can't include:
\itemize{
\item \code{func}
\item \code{args}
\item \code{libpath}
}}
}
\value{
A \verb{<bench_mark>}.
}
\description{
\code{bench()} allows you to run a single function, \code{fn}, multiple times in
separate R sessions, where each R session has different versions of packages
installed. A typical use case is running a before/after benchmark, comparing
the CRAN version of a package with a development version of the same package.

For example, \code{bench(fn, pkgs = c("vctrs", "r-lib/vctrs#100"))} would run \code{fn}
in 2 separate R sessions, one with CRAN vctrs installed, and one with the
pull request installed.

When using \code{bench()}, each call to \code{fn} must return a \verb{<bench_mark>} data
frame from the bench package. This is typically generated using either a
call to \code{\link[bench:mark]{bench::mark()}} or \code{\link[bench:press]{bench::press()}}.

\code{bench()} is similar to \code{\link[=run]{run()}}, but is specifically designed to be useful in
conjunction with the bench package when benchmarking across multiple package
versions.
}
\details{
The installed packages are placed in temporary directories that act as an
extra library path. The temporary directory is prepended to \code{libpath}, which
is then passed along to \code{\link[callr:r]{callr::r()}} as the \code{libpath} argument there. This
ensures that your personal package libraries remain untouched.
}
\section{Global options}{

\itemize{
\item \code{cross.quiet}: \verb{[TRUE / FALSE]}

Should cross specific messages about the process be shown?

Defaults to \code{!rlang::is_interactive()} if not set, so messages are only
shown when running interactively.
\item \code{cross.pak_quiet}: \verb{[TRUE / FALSE]}

Should pak installation messages be shown?

Defaults to \code{TRUE} if not set.
}
}

\examples{
\dontshow{if (FALSE) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
# Run a benchmark across 2 different versions of vctrs
bench(pkgs = c("vctrs", "r-lib/vctrs"), ~{
  library(vctrs)
  x <- c(TRUE, FALSE, NA, TRUE)
  bench::mark(vec_detect_missing(x))
})

# You can also use `bench::press()` to generate a grid
bench(pkgs = c("vctrs", "r-lib/vctrs"), ~{
  library(vctrs)
  set.seed(123)
  x <- sample(100)

  bench::press(
    table = c(10, 20, 30),
    {
      table <- seq_len(table)
      bench::mark(vec_match(x, table))
    }
  )
})
\dontshow{\}) # examplesIf}
}
