% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cross-package.R
\docType{package}
\name{cross-package}
\alias{cross}
\alias{cross-package}
\title{cross: Run Functions Across Package Versions}
\description{
Providings tooling to run a single function across multiple package versions, which is particularly useful for running before and after benchmarks.
}
\author{
\strong{Maintainer}: Davis Vaughan \email{davis@posit.co}

Other contributors:
\itemize{
  \item Posit Software, PBC [copyright holder, funder]
}

}
\keyword{internal}
