% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run.R
\name{run_versions}
\alias{run_versions}
\title{Evaluate an expression across different package versions}
\usage{
run_versions(
  expr,
  ...,
  pkgs,
  libpath = .libPaths(),
  args_pak = list(),
  args_callr = list()
)
}
\arguments{
\item{expr}{\verb{[expression]}

An expression to evaluate. The expression is passed along to \code{\link[callr:r]{callr::r()}}
as the body of a function with zero arguments, so it is evaluated in a
fresh R session and must be self-contained.

Read the \code{func} docs of \code{\link[callr:r]{callr::r()}} for the full set of restrictions on
\code{expr}.}

\item{...}{These dots are for future extensions and must be empty.}

\item{pkgs}{\verb{[character]}

A character vector of package names or remote package specifications to
install and run \code{expr} against. Passed along to \code{\link[pak:pkg_install]{pak::pkg_install()}}.}

\item{libpath}{\verb{[character]}

The base library path to use.}

\item{args_pak}{\verb{[named list]}

A named list of arguments to pass on to \code{\link[pak:pkg_install]{pak::pkg_install()}}.

Can't include:
\itemize{
\item \code{pkg}
\item \code{lib}
\item \code{ask}
}}

\item{args_callr}{\verb{[named list]}

A named list of arguments to pass on to \code{\link[callr:r]{callr::r()}}.

Can't include:
\itemize{
\item \code{func}
\item \code{args}
\item \code{libpath}
}}
}
\value{
A data frame with two columns:
\itemize{
\item \code{pkg}, a character vector containing \code{pkgs}.
\item \code{result}, a list column containing the result of calling \code{expr} for that
version of the package.
}
}
\description{
\code{run_versions()} allows you to run a single expression, \code{expr}, multiple
times in separate R sessions, where each R sessions has different versions of
packages installed. If you're looking to run benchmarks across different
versions of packages, you likely want \code{\link[=bench_versions]{bench_versions()}} instead.

For example, \code{run_versions(expr, pkgs = c("vctrs", "r-lib/vctrs#100"))} would
run \code{expr} in 2 separate R sessions, one with CRAN vctrs installed, and one
with the pull request installed.
}
\details{
The installed packages are placed in temporary directories that act as an
extra library path. The temporary directory is prepended to \code{libpath}, which
is then passed along to \code{\link[callr:r]{callr::r()}} as the \code{libpath} argument there. This
ensures that your personal package libraries remain untouched.
}
\section{Global options}{

\itemize{
\item \code{cross.quiet}: \verb{[TRUE / FALSE]}

Should cross specific messages about the process be shown?

Defaults to \code{!rlang::is_interactive()} if not set, so messages are only
shown when running interactively.
\item \code{cross.pak_quiet}: \verb{[TRUE / FALSE]}

Should pak installation messages be shown?

Defaults to \code{TRUE} if not set.
}
}

\examples{
\dontshow{if (FALSE) withAutoprint(\{ # examplesIf}
# Run a benchmark across 2 different versions of vctrs
# (See `bench_versions()` for an even easier way)
run_versions(pkgs = c("vctrs", "r-lib/vctrs"), {
  library(vctrs)
  x <- c(TRUE, FALSE, NA, TRUE)
  bench::mark(vec_detect_missing(x))
})
\dontshow{\}) # examplesIf}
}
