% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cross-package.R
\docType{package}
\name{cross-package}
\alias{cross}
\alias{cross-package}
\title{cross: Run Functions Across Package Versions}
\description{
Tooling to run a single function across multiple package versions or across multiple local branches, which is particularly useful for running before and after benchmarks.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/DavisVaughan/cross}
  \item Report bugs at \url{https://github.com/DavisVaughan/cross/issues}
}

}
\author{
\strong{Maintainer}: Davis Vaughan \email{davis@posit.co}

Other contributors:
\itemize{
  \item Posit Software, PBC [copyright holder, funder]
}

}
\keyword{internal}
